Fixed issues with inconsistent dataType

+++ src/main/java/com/dsny/hack/service/EventProcessorImpl.java
@@ -1,7 +1,5 @@
 package com.dsny.hack.service;
 
-import java.text.SimpleDateFormat;
-import java.util.Date;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
@@ -35,11 +33,11 @@ public class EventProcessorImpl implements EventProcessor {
 	public void processNewEvents() {
 		List<Event> events = eventRepo.findByStatus("New");
 		// group by transactionID
-		Map<Long, List<Event>> eventMap = events.stream().collect(
+		Map<Integer, List<Event>> eventMap = events.stream().collect(
 				Collectors.groupingBy(Event::getTransactionID));
 
 		if (eventMap.size() > 0) {
-			for (Entry<Long, List<Event>> entry : eventMap.entrySet()) {
+			for (Entry<Integer, List<Event>> entry : eventMap.entrySet()) {
 				boolean isAllEventReady = processEvents(entry.getValue());
 				if (isAllEventReady) {
 					entry.getValue().forEach((e) -> e.setStatus("Completed"));
